Gros commit après inactivité

+++ src/main/java/net/onima/onimaapi/mod/items/CompassBooster.java
@@ -34,7 +34,7 @@ public class CompassBooster extends ModItem implements EntityClickable { //TODO 
 			return;
 		}
 		
-		Methods.boost(target, player.getOptions().<Double>get(PlayerOption.ModOptions.ENTITY_BOOSTER), player.getOptions().<Double>get(PlayerOption.ModOptions.ENTITY_BOOSTER), player.getOptions().<Float>get(PlayerOption.ModOptions.ENTITY_BOOSTER));
+		Methods.boost(target, player.getOptions().<Double>get(PlayerOption.ModOptions.ENTITY_BOOSTER), player.getOptions().<Double>get(PlayerOption.ModOptions.ENTITY_BOOSTER), player.getOptions().<Double>get(PlayerOption.ModOptions.ENTITY_BOOSTER));
 		
 		player.sendMessage(MOD_PREFIX + " §fWOOOOOOOSSSSHHH !!!!");
 		boostSound.play(player);
